src/frontend/requirements.txt
Flask==2.3.2
python-dotenv
clickhouse-driver
prefect==2.10.4
pandas
pyarrow
clickhouse-connect
streamlit
plotly